backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
APScheduler==3.10.4
asyncpg==0.31.0
attrs==25.4.0
beautifulsoup4==4.12.3
black==25.11.0
bleach==6.3.0
blinker==1.9.0
CacheControl==0.14.4
cachetools==6.2.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
deep-translator==1.11.4
deprecation==2.1.0
et_xmlfile==2.0.0
fake-useragent==1.5.1
fastapi==0.123.8
filelock==3.20.0
firebase-admin==6.3.0
Flask==3.1.2
fonttools==4.60.1
fsspec==2025.10.0
fuzzywuzzy==0.18.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.119.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-cloud-core==2.3.3
google-cloud-firestore==2.14.0
google-cloud-storage==3.0.0
google-cloud-vision==3.11.0
google-crc32c==1.7.1
google-generativeai==0.8.5
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.62.3
gTTS==2.5.4
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.27.0
hyperframe==6.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kaggle==1.8.0
kagglesdk==0.1.13
kiwisolver==1.4.9
Levenshtein==0.27.3
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
mypy==1.18.2
mypy_extensions==1.1.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
polars==1.35.2
polars-runtime-32==1.35.2
postgrest==2.25.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==4.25.2
psutil==7.1.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyJWT==2.10.1
pyparsing==3.2.5
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.27.3
python-multipart==0.0.21
python-slugify==8.0.4
pytokens==0.3.0
pytz==2025.2
PyYAML==6.0.3
RapidFuzz==3.14.3
realtime==2.25.0
requests==2.31.0
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.16.3
seaborn==0.13.2
selenium==4.38.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
starlette==0.50.0
storage3==2.25.0
StrEnum==0.4.15
supabase==2.25.0
supabase-auth==2.25.0
supabase-functions==2.25.0
sympy==1.14.0
text-unidecode==1.3
threadpoolctl==3.6.0
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
trio==0.32.0
trio-websocket==0.12.2
types-requests==2.32.4.20250913
types-tqdm==4.67.0.20250809
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
ultralytics==8.3.230
ultralytics-thop==2.0.18
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.27.0
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.3.2
xgboost==3.1.2
yarl==1.22.0
